Employees CRUD
The admin is now able to add new employees to the database and update an existing employee. Fixed an issue when updating an existing employees details would make a new entry. The update and create functions now work as intended. Fixed multiple issues with the routes files looking for a none existing route. The admin user is now able to use full CRUD functionality on the employees and companies tables.