Table layout fixes.
Added the pagination for the companies and employees listing views to be part of the table layout. Added view, edit and delete buttons to the table layout of employees on the company show view. Changed the layout of the back to listing button and the update button on the companies show view to be more responsive.