Dashboard rework
The dashboard now contains employee and company cards that shows how many employees and companies are registered within the system. Implemented a most recent card for the company and employees sections. These cards show the 5 most recent company and employee entries with buttons to view, edit or delete these entries from the system.